Editing and Deleting finally working properly

+++ app/src/main/java/com/artemthedev/simpleandroidtaskmanager/MainActivity.java
@@ -1,19 +1,17 @@
 package com.artemthedev.simpleandroidtaskmanager;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
-
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
+import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -25,7 +23,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        if (getSupportFragmentManager().getFragments().size() == 0) displayNotes();
     }
 
     @Override
@@ -33,6 +30,12 @@ public class MainActivity extends AppCompatActivity {
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.menu, menu);
         return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        displayNotes();
     }
 
     @Override
@@ -55,10 +58,36 @@ public class MainActivity extends AppCompatActivity {
         startActivity(intent);
     }
 
-    public void clearDB(MenuItem item) {
-        openHelper = new NotesDBHelper(this);
-        db = openHelper.getReadableDatabase();
-        db.delete("NOTES", null, null);
-        displayNotes();
+    public void clearNotes(MenuItem item) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setMessage(R.string.erasing_notes_warning)
+                .setTitle(R.string.warning);
+        builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
+            public void onClick(DialogInterface dialog, int id) {
+                if (!fragment.isEmpty) {
+                    openHelper = new NotesDBHelper(MainActivity.this);
+                    db = openHelper.getReadableDatabase();
+                    db.delete("NOTES", null, null);
+                    Toast.makeText(MainActivity.this, R.string.db_notes_cleaned, Toast.LENGTH_SHORT).show();
+                    onResume();
+                }
+            }
+        });
+        builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
+            public void onClick(DialogInterface dialog, int id) {}
+        });
+        AlertDialog dialog = builder.create();
+        dialog.show();
+    }
+
+    public void toSettings(MenuItem item) {
+        Intent intent = new Intent(this, SettingsActivity.class);
+        startActivity(intent);
+    }
+
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
+        super.onBackPressed();
     }
 }
